requirements.txt
pynetbox==7.2.0
python-dotenv==1.0.1 